Advance on data source list component. Creation of mixin for column-content
