+++ client/app/app.constants.js
@@ -2,7 +2,7 @@
 module.exports = angular.module("app.constants", [])
 
 .constant("Properties", {
-	"endpoint": "http://10.164.97.163:8080/api"
+	"endpoint": "http://localhost:8889/api"
 })
 
 ;
