+++ client/app/common/mixins/_mixins.scss
@@ -5,3 +5,4 @@
 @import "link";
 @import "spin";
 @import "transitions";
+@import "column-content";
