+++ client/app/common/modal/modal.scss
@@ -25,7 +25,7 @@ ual-modal.-fullmodal {
     background:  rgba($primaryWhiteColor, 0.9);
     > ng-transclude {
         height: 100%;
-        max-height: 100%;
+        // max-height: 100%;
         @include gridle(gutter 8% 0px);
         @include gridle( row row-align center);
         > * {
