+++ client/app/components/dashboard/ualReport/ualDataSource/ualDataSource.html
@@ -13,7 +13,7 @@
         </div>
 
         <div class="buttons">
-            <button class="apply">Apply</button>
+            <ual-button class="-primary">Apply</ual-button>
         </div>
     </div>
 </ual-data-source>
