+++ client/app/components/dashboard/ualReport/ualDataSource/ualDataSource.js
@@ -1,7 +1,9 @@
 import angular from 'angular';
 import ualDataSourceFactory from './ualDataSource.factory';
+import ualDataSourceList from './ualDataSourceList/ualDataSourceList';
 
 let ualDataSourceModule = angular.module('ualDataSource', [
+    ualDataSourceList.name
 ])
 
 .factory('ualDataSource', ualDataSourceFactory);
