+++ client/app/components/dashboard/ualReport/ualDataSource/ualDataSourceList/ualDataSourceList.html
@@ -1,0 +1,8 @@
+<div ng-repeat="group in vm.groupList">
+    <div class="group-name">
+        {{group.name}}
+    </div>
+    <div ng-repeat-start="item in group.elements" data-id="{{item.id}}" class="group-item"  ng-repeat-end="">
+        {{item.name}}
+    </div>
+</div>
