--- client/app/components/dashboard/ualReport/ualDataSourceList/ualDataSourceList.html
@@ -1,11 +1,0 @@
-<div ng-repeat="group in vm.groupList" >
-    <div class="group-name">
-        {{group.name}}
-    </div>
-    <div ng-repeat="item in group.elements">
-        <div data-id="{{item.id}}" class="group-item">
-            {{item.name}}
-        </div>
-    </div>
-
-</div>
