--- client/app/components/dashboard/ualReport/ualDataSourceList/ualDataSourceList.scss
@@ -1,20 +1,0 @@
-@import '../../../../common/common';
-
-ual-data-source-list { 
-  @include gridle(col col-align left);
-    -webkit-column-count: 3;
-    -moz-column-count: 3;
-        column-count: 3;
-  .group-name{
-      display: inline-block;
-      @include gridle(4);
-      @include font-clanot-bold(20);
-      @include gridle(gutter 24px 0px)
-  }
-  .group-item{
-      display: inline-block;
-      @include gridle(4);
-      @include gridle(gutter 10px 0px);
-      @include font-clanot-book(18);
-  }
-}
