+++ client/app/components/dashboard/ualReport/ualReport.js
@@ -1,14 +1,12 @@
 import angular from 'angular';
-import ualReportList from './ualReportList/ualReportList'
-import ualReportForm from './ualReportForm/ualReportForm'
-import ualDataSourceList from './ualDataSourceList/ualDataSourceList';
+import ualReportList from './ualReportList/ualReportList';
+import ualReportForm from './ualReportForm/ualReportForm';
 import ualDataSource from './ualDataSource/ualDataSource';
 
 let ualReportModule = angular.module('ualReport', [
     ualDataSource.name,
     ualReportList.name,
-    ualReportForm.name,
-    ualDataSourceList.name
+    ualReportForm.name
 ])
 
 export default ualReportModule;
