+++ client/app/components/dashboard/ualReport/ualReportForm/ualReportForm.html
@@ -1,7 +1,9 @@
 <div class="ual-report-form">
     <div>
         <div class="appName">Create Report</div>
-        <div><ual-button class="-secondary">Save</ual-button></div>
+        <div>
+            <ual-button class="-secondary">Save</ual-button>
+        </div>
     </div>
     <div class="reportName separatorLine">
         <div>
@@ -9,7 +11,11 @@
         </div>
     </div>
     <div class="info separatorLine">
-        <div><span>Data Source:</span> <b>Polaris</b> <a ng-click="vm.openDataSourceModal()">Change Data Source</a></div>
+        <div>
+            <span>Data Source:</span>
+            <b>Polaris</b>
+            <a ng-click="vm.openDataSourceModal()">Change Data Source</a>
+        </div>
         <div><span>Variables:</span> <b>10 Variables Selected</b> <a href="#">Manage Variables</a></div>
     </div>
     <div class="info">
